+++ CHANGES.md
@@ -1,0 +1,8 @@
+## v2018.01.07
+
+[screenshot][screenshot-2018-01-07] | [demo permalink][demo-2018-01-07]
+
+- Mouse up and down to show zooming into random nodes of the tree
+
+[screenshot-2018-01-07]:https://user-images.githubusercontent.com/116838/34665780-2b68bb38-f427-11e7-96c1-95c80ed3f39c.gif
+[demo-2018-01-07]:http://btrdb-viz-2018-01-07.surge.sh/
