+++ src/index.css
@@ -1,3 +1,11 @@
+/* Allow all elements to fill page */
+html,
+body,
+#root {
+  width: 100%;
+  height: 100%;
+}
+
 body {
   margin: 0;
   padding: 0;
